Prevent request to undefined.undefined Fixes #33
Also do not display a broken image until it loads, conditionaly show the image tag

Also found out that it was calling random image every render, and now twice with the conditional check. useMemo was the answer and what I expected :)

Fixed useEffect shenanigan where it would try to load up a currentSongHash of undefined and set that as the current song which had a trickle down of issues.